Refactor P619 model  - Create separate classes for reference atmosphere and scintillation
