+++ tests/test_propagation_p619.py
@@ -10,35 +10,22 @@ import numpy as np
 import numpy.testing as npt
 from sharc.propagation.propagation_p619 import PropagationP619
 
+
+class DummyParams():
+    def __init__(self):
+        self.imt_altitude = 0
+        self.surf_water_vapour_density = 0
+        self.EARTH_RADIUS = 6371000
+
 class TestPropagationP619(unittest.TestCase):
 
     def setUp(self):
         self.p619 = PropagationP619()
 
-    def test_specific_attenuation(self):
-        temperature = 15 + 273.15 # K
-        vapour_density = 7.5 # g/m**3
-        pressure_hPa = 1013.25
-        vapour_pressure_hPa = vapour_density * temperature / 216.7
-
-        # compare with benchmark from ITU-R P-676-11 Figs. 1 and 2
-        f_GHz_vec = [50, 60, 100, 200, 500, 1000]
-        specific_att = np.zeros(len(f_GHz_vec))
-        specific_att_p676_lower = [3e-1, 1e1, 4e-1, 2, 5e1, 6e2]
-        specific_att_p676_upper = [5e-1, 2e1, 5e-1, 4, 7e1, 8e2]
-
-        for index in range(len(f_GHz_vec)):
-            specific_att[index] = self.p619._get_specific_attenuation(pressure_hPa,
-                                                                      vapour_pressure_hPa,
-                                                                      temperature,
-                                                                      float(f_GHz_vec[index]) * 1000)
-        npt.assert_array_less(specific_att_p676_lower, specific_att)
-        npt.assert_array_less(specific_att, specific_att_p676_upper)
-
     def test_atmospheric_gasses_loss (self):
         # compare with benchmark from ITU-R P-619 Fig. 3
         frequency_MHz = 30000.
-        sat_params = ParametersFss()
+        sat_params = DummyParams()
         sat_params.imt_altitude = 1000
 
         vapour_density_vec = [7.5, 12.5, 2.5, 5.]
@@ -52,8 +39,10 @@ class TestPropagationP619(unittest.TestCase):
                                                                               loss_lower_vec,
                                                                               loss_upper_vec):
             sat_params.surf_water_vapour_density = vapour_density
-            loss = self.p619._get_atmospheric_gasses_loss(frequency_MHz, apparent_elevation,
-                                                          sat_params)
+            loss = self.p619._get_atmospheric_gasses_loss(frequency_MHz=frequency_MHz,
+                                                          apparent_elevation=apparent_elevation,
+                                                          surf_water_vapour_density=vapour_density,
+                                                          sat_params=sat_params)
             self.assertLessEqual(loss_lower, loss)
             self.assertGreaterEqual(loss_upper, loss)
 
@@ -82,28 +71,6 @@ class TestPropagationP619(unittest.TestCase):
             else:
                 self.assertLess(attenuation, 0)
 
-    def test_tropo_scintillation_attenuation(self):
-        # compare with benchmark from ITU-R P-619 Fig. 8
-
-        antenna_gain = 0.
-        frequency_MHz = 30000.
-        wet_refractivity = 42.5
-
-        elevation_vec = np.array([5., 10., 20., 90., 35., 5., 10., 20., 35., 90.])
-        percentage_gain_exceeded = np.array([.01, .1, 1, 3, 10, 90, 98, 99, 99.9, 99.99])
-        attenuation_lower = [5, 1, .5, .1, .1, 1, 1, .5, .4, .3]
-        attenuation_upper = [7, 2, .6, .2, .2, 2, 2, .7, .6, .5]
-        sign = [-1, -1, -1, -1, -1, +1, +1, +1, +1, +1]
-        attenuation = self.p619._get_tropospheric_scintillation(elevation=elevation_vec,
-                                                                frequency_MHz=frequency_MHz,
-                                                                antenna_gain_dB=antenna_gain,
-                                                                time_ratio=percentage_gain_exceeded / 100,
-                                                                wet_refractivity=wet_refractivity)
-
-        npt.assert_array_less(attenuation_lower, np.abs(attenuation))
-        npt.assert_array_less(np.abs(attenuation), attenuation_upper)
-
-        npt.assert_array_equal(np.sign(attenuation), sign)
 
 if __name__ == '__main__':
     unittest.main()
